backend/notes/serializers.py
from rest_framework import serializers

from .models import Note, Tag, Category


class TagListSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = ['id', 'name']


class CategoryListSerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'name']


class NoteListSerializer(serializers.ModelSerializer):
    tags = TagListSerializer(many=True)
    categories = CategoryListSerializer(many=True)

    class Meta:
        model = Note
        fields = ['id', 'title', 'content', 'tags', 'categories', 'archived']


class NoteSerializer(serializers.ModelSerializer):
    tags = serializers.PrimaryKeyRelatedField(many=True, queryset=Tag.objects.all())
    categories = serializers.PrimaryKeyRelatedField(many=True, queryset=Category.objects.all())

    class Meta:
        model = Note
        fields = ['id', 'title', 'content', 'tags', 'categories', 'archived']
