backend/requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
python-decouple==3.8
sqlparse==0.4.4
